updated database structure, created stock routes

+++ db/migrations/20200518191717-create-company.js
@@ -41,11 +41,11 @@ module.exports = {
         allowNull: false
       },
       priceEarningRatio: {
-        type: Sequelize.FLOAT(6,2),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       dividendYield: {
-        type: Sequelize.FLOAT(6,2),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       averageVolume: {
